+++ src/vs/workbench/browser/parts/activitybar/media/activitybarpart.css
@@ -7,15 +7,10 @@
 	width: 48px;
 }
 
-.monaco-workbench .part > .drag-overlay {
-	transition-property: opacity;
-	transition-duration: .2s;
-	width: 100%;
-	height: 100%;
-	position: absolute;
-	opacity: 0;
-	top: 0;
-	pointer-events: none;
+.monaco-workbench .part > .drop-block-overlay.visible {
+	visibility: visible;
+	backdrop-filter: brightness(97%) blur(2px);
+	opacity: 1;
 }
 
 .monaco-workbench .activitybar > .content {
